chore: integrate `schema-tools-versioning`
