+++ package.json
@@ -18,17 +18,17 @@
     "lint:staged": "lint-staged",
     "format": "prettier --write \"**/*.{js,json}\"",
     "test": "jest --runInBand",
-    "update-version": "node ./update-version",
+    "update-version": "node ./utils/update-version",
     "release": "read -p 'GITHUB_TOKEN: ' GITHUB_TOKEN && export GITHUB_TOKEN=$GITHUB_TOKEN && release-it"
   },
   "devDependencies": {
-    "@s1seven/schema-tools-utils": "^0.0.20",
+    "@s1seven/schema-tools-utils": "0.0.25",
+    "@s1seven/schema-tools-versioning": "^0.1.0",
     "ajv": "^8.8.0",
     "ajv-formats": "^2.1.0",
     "eslint": "^8.4.0",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-prettier": "^4.0.0",
-    "glob": "^7.2.0",
     "jest": "^27.4.3",
     "lint-staged": "^12.1.2",
     "pre-commit": "^1.2.2",
