+++ utils/constants.js
@@ -1,0 +1,5 @@
+const defaultServerUrl = 'https://schemas.s1seven.com/e-coc-schemas';
+
+module.exports = {
+  defaultServerUrl,
+};
